move tests to tests.rs
