+++ xtc/src/cycles_wallet.rs
@@ -16,28 +16,28 @@ use ic_kit::{get_context, Context, Principal};
 use serde::*;
 
 #[query]
-fn name() -> Option<&'static str> {
+pub fn name() -> Option<&'static str> {
     Some(meta().name)
 }
 
 #[derive(CandidType, Deserialize)]
-struct CallCanisterArgs {
-    canister: Principal,
-    method_name: String,
+pub struct CallCanisterArgs {
+    pub canister: Principal,
+    pub method_name: String,
     #[serde(with = "serde_bytes")]
-    args: Vec<u8>,
-    cycles: u64,
+    pub args: Vec<u8>,
+    pub cycles: u64,
 }
 
 #[derive(CandidType, Deserialize)]
-struct CallResult {
+pub struct CallResult {
     #[serde(with = "serde_bytes")]
-    r#return: Vec<u8>,
+    pub r#return: Vec<u8>,
 }
 
 /// Forward a call to another canister.
 #[update(name = "wallet_call")]
-async fn call(args: CallCanisterArgs) -> Result<CallResult, String> {
+pub async fn call(args: CallCanisterArgs) -> Result<CallResult, String> {
     IsShutDown::guard();
 
     let ic = get_context();
@@ -96,13 +96,13 @@ async fn call(args: CallCanisterArgs) -> Result<CallResult, String> {
 // Create canister call
 
 #[derive(CandidType, Deserialize)]
-struct CreateCanisterArgs {
-    cycles: u64,
-    controller: Option<Principal>,
+pub struct CreateCanisterArgs {
+    pub cycles: u64,
+    pub controller: Option<Principal>,
 }
 
 #[update(name = "wallet_create_canister")]
-async fn create_canister(args: CreateCanisterArgs) -> Result<WithCanisterId, String> {
+pub async fn create_canister(args: CreateCanisterArgs) -> Result<WithCanisterId, String> {
     IsShutDown::guard();
 
     let ic = get_context();
@@ -167,12 +167,12 @@ async fn create_canister(args: CreateCanisterArgs) -> Result<WithCanisterId, Str
 }
 
 #[derive(CandidType)]
-struct BalanceResult {
-    amount: u64,
+pub struct BalanceResult {
+    pub amount: u64,
 }
 
 #[query]
-fn wallet_balance() -> BalanceResult {
+pub fn wallet_balance() -> BalanceResult {
     let ic = get_context();
     let ledger = ic.get::<Ledger>();
     let amount = ledger.balance(&ic.caller());
@@ -180,13 +180,13 @@ fn wallet_balance() -> BalanceResult {
 }
 
 #[derive(CandidType, Deserialize)]
-struct SendCyclesArgs {
-    canister: Principal,
-    amount: u64,
+pub struct SendCyclesArgs {
+    pub canister: Principal,
+    pub amount: u64,
 }
 
 #[update]
-async fn wallet_send(args: SendCyclesArgs) -> Result<(), String> {
+pub async fn wallet_send(args: SendCyclesArgs) -> Result<(), String> {
     IsShutDown::guard();
 
     let ic = get_context();
@@ -242,139 +242,10 @@ async fn wallet_send(args: SendCyclesArgs) -> Result<(), String> {
 }
 
 #[update]
-async fn wallet_create_wallet(_: CreateCanisterArgs) -> Result<WithCanisterId, String> {
+pub async fn wallet_create_wallet(_: CreateCanisterArgs) -> Result<WithCanisterId, String> {
     let ic = get_context();
     crate::progress().await;
     Ok(WithCanisterId {
         canister_id: ic.id(),
     })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::ledger::Ledger;
-    use ic_kit::{async_test, Context, MockContext, RejectionCode};
-    use ic_kit::{mock_principals, Method, RawHandler};
-    use std::fmt::Debug;
-    use std::future::Future;
-    use std::pin::Pin;
-
-    fn reset_ledger(ctx: &mut MockContext) {
-        let ledger = ctx.get_mut::<Ledger>();
-        ledger.deposit(Principal::anonymous(), 3 * 10_000_000_000_000);
-        ledger.load(vec![
-            (mock_principals::alice(), 10_000_000_000_000),
-            (mock_principals::bob(), 10_000_000_000_000),
-            (mock_principals::john(), 10_000_000_000_000),
-        ]);
-    }
-
-    /// General method to test the behaviour of fee implementation of methods.
-    async fn test_fee<T: CandidType + Clone, O, E: Debug>(
-        response: T,
-        cb: Box<dyn Fn(u64) -> Pin<Box<dyn Future<Output = Result<O, E>>>>>,
-    ) {
-        let ctx = MockContext::new()
-            .with_caller(mock_principals::alice())
-            .inject();
-
-        // Consumes all
-        reset_ledger(ctx);
-        ctx.clear_handlers();
-        ctx.use_handler(
-            Method::new()
-                .response(response.clone())
-                .cycles_consume(2_000),
-        );
-        cb(1_000).await.expect("Unexpected error.");
-        ctx.call_state_reset();
-
-        assert_eq!(
-            ctx.get::<Ledger>().balance(&mock_principals::alice()),
-            10_000_000_000_000 - 1_000 - compute_fee(1_000)
-        );
-
-        // With refund.
-        reset_ledger(ctx);
-        cb(10_000).await.expect("Unexpected error.");
-        ctx.call_state_reset();
-
-        assert_eq!(
-            ctx.get::<Ledger>().balance(&mock_principals::alice()),
-            10_000_000_000_000 - 2_000 - compute_fee(2_000)
-        );
-
-        // With error.
-        reset_ledger(ctx);
-        ctx.clear_handlers();
-        ctx.use_handler(RawHandler::raw(Box::new(|_, _, _, _| {
-            Err((
-                RejectionCode::DestinationInvalid,
-                "Canister not found.".into(),
-            ))
-        })));
-        cb(10_000).await.err().expect("Expected Err response.");
-        ctx.call_state_reset();
-
-        assert_eq!(
-            ctx.get::<Ledger>().balance(&mock_principals::alice()),
-            10_000_000_000_000
-        );
-    }
-
-    #[async_test]
-    async fn wallet_call_fee() {
-        test_fee(
-            (),
-            Box::new(|cycles| {
-                Box::pin(async move {
-                    call(CallCanisterArgs {
-                        canister: mock_principals::john(),
-                        method_name: "xxx".to_string(),
-                        args: vec![],
-                        cycles,
-                    })
-                    .await
-                })
-            }),
-        )
-        .await;
-    }
-
-    #[async_test]
-    async fn create_canister_fee() {
-        test_fee(
-            WithCanisterId {
-                canister_id: mock_principals::xtc(),
-            },
-            Box::new(|cycles| {
-                Box::pin(async move {
-                    create_canister(CreateCanisterArgs {
-                        cycles,
-                        controller: None,
-                    })
-                    .await
-                })
-            }),
-        )
-        .await;
-    }
-
-    #[async_test]
-    async fn send_fee() {
-        test_fee(
-            (),
-            Box::new(|cycles| {
-                Box::pin(async move {
-                    wallet_send(SendCyclesArgs {
-                        canister: mock_principals::xtc(),
-                        amount: cycles,
-                    })
-                    .await
-                })
-            }),
-        )
-        .await;
-    }
 }
