+++ xtc/src/lib.rs
@@ -7,6 +7,9 @@ mod meta;
 mod stats;
 mod upgrade;
 
+#[cfg(test)]
+mod tests;
+
 /// Perform only one pending async task, returns whether an async call was performed
 /// as the result of calling this method or not.
 /// This method should only be called from updates.
